fix: change atoi to atol for better handling of large numbers in primes computation

+++ mpi/primes.cpp
@@ -97,7 +97,7 @@ int main(int argc, char **argv) {
     MPI_Comm_size(MPI_COMM_WORLD, &size);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
-    long n = atoi(argv[1]);
+    long n = atol(argv[1]);
 
     if (rank == ROOT_RANK) {
         master(n, size);
